# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  dto.MoveTicketRequest:
    properties:
      new_board_id:
        type: integer
      ticket_id:
        type: integer
    type: object
  dto.MoveTicketResponse:
    properties:
      message:
        type: string
    type: object
  dto.SearchTaskResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      label_id:
        type: integer
      name:
        type: string
      ticket_id:
        type: integer
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: iiaksisii@gmail.com
    name: Антон
    url: https://github.com/IIAkSISII
  description: Разработка бекенда для отслеживания задач
  title: Трекер задач
  version: "1.0"
paths:
...
  /task/search:
    get:
      consumes:
      - application/json
      description: Поиск задач по названию, метке и пользователю
      parameters:
      - description: Часть названия задачи
        in: query
        name: name
        type: string
      - description: ID метки
        in: query
        name: label_id
        type: integer
      - description: ID исполнителя
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список найденных задач
          schema:
            items:
              $ref: '#/definitions/dto.SearchTaskResponse'
            type: array
        "400":
          description: Invalid label_id or user_id
          schema:
            type: string
        "500":
          description: 'Failed to search tasks: <описание ошибки>'
          schema:
            type: string
      summary: Поиск задач
      tags:
      - Tasks
...
  /ticket/move:
    put:
      consumes:
      - application/json
      description: |-
        Перемещаем карточку с указанием id карточки и id новой доски.
        Параметры передаются в теле запроса в качестве json-объекта.
        Если карточку не удается создать, возвращаем ошибку.
      parameters:
      - description: ticket_id, new_board_id
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.MoveTicketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Карточка успешно перемещена
          schema:
            $ref: '#/definitions/dto.MoveTicketResponse'
        "400":
          description: Invalid json payload, Name is required или TicketId or BoardId
            must be positive
          schema:
            type: string
        "500":
          description: 'Cannot move ticket: <описание ошибки>'
          schema:
            type: string
      summary: Переместить карточку на другую доску
      tags:
      - Tickets
...
produces:
- json text/plain
schemes:
- http
- https
swagger: "2.0"